Potions added
Potions added:
Health
Speed
Attack Speed
Damage

Attack speed is not functional yet; the Weapon class's 'speed' value doesn't seem to be fully implemented.

All potions (except health) trigger particle effects on the player for their duration so the player knows when it has worn off. This could be quite easily expanded so that the player is given a warning a few seconds before as well.