Room Management Added
Rooms object contains a script called RoomManager which contains a method called 'getCurrentRoom()' which returns an Room object (which is a child of Rooms), which contains a script called  FogOfWar. This script calculates and stores data on the size of the room and whether or not it is a corridoor.  It also allows for exploration of the dungeon by applying a dark filter to the rooms which the player has not yet entered. Any enemies within the filter are disabled and made invisible.

I though this would take me ~4 hours.
It took all weekend...