Player death implemented with cool fades
More pain than it was probably worth...

PlayerDeath.cs was created and added to the Player 1 prefab.
It has 1 inspector value - a PlayerCombat.cs reference.

Fade.cs was modified slightly to add implementation I should have added originally...

A new GameObject was added to the scene & created as a UI prefab: DeathImage.
This is a UI image with a Fade.cs script. The image file itself is in Assets/Art/UI. The Image component of the object must be disabled by default.